+++ CHANGELOG.md
@@ -4,7 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.11.0-rc.1] - Unreleased
+## [0.11.0] - Unreleased
+### Added
+- Allow passing an explicit credentials secret to the device instead of depending from the on-board
+  agent registration.
+- Add `astarte_device_is_connected` function.
+
+## [0.11.0-rc.1] - 2020-03-26
 ### Changed
 - Deprecate implicit credentials initialization, the user should now call `astarte_credentials_init`
   in its `main` before calling `astarte_device_init`.